feature: add ortho

+++ Shader.cpp
@@ -2,7 +2,7 @@
 #include "mgl.h"
 
 int Shader::uniform(const float* el, int size, int location) {
-    if(location >= uniforms_offset.size()) {
+    if(location >= (int)uniforms_offset.size()) {
         return -1;
     }
     if(location < 0) {
@@ -10,7 +10,7 @@ int Shader::uniform(const float* el, int size, int location) {
         int pre = 0;
         if(!uniforms_offset.empty()) pre = uniforms_offset.back();
         uniforms_offset.emplace_back(size + pre);
-        return uniforms_offset.size() - 1;
+        return (int)uniforms_offset.size() - 1;
     }
     int pre = location > 0 ? uniforms_offset[location - 1] : 0;
     int cur = uniforms_offset[location];
@@ -30,7 +30,7 @@ const float* Shader::getvaring(floatstream& varying, int size, int& offset) {
 }
 
 const float* Shader::getunif(int location, int check) const {
-    if(location < 0 || location >= uniforms.size()) return NULL;
+    if(location < 0 || location >= (int)uniforms.size()) return NULL;
     int offset = 0;
     if(location) offset = uniforms_offset[location - 1];
     if(check >= 0 && check != uniforms_offset[location] - offset) return NULL;
